+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 This file should follow the standards specified on [http://keepachangelog.com/]
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleaed]
+
+### Added
+
+- Optional three-argument signature for `ActiveRel#create` and `#create!`, just like `initialize`.
+
 ## [6.0.0.alpha.11] - 11-3-2015
 
 ### Fixed